mole/pyproject.toml
[project]
name = "mole-browser"
version = "0.1.0"
description = "Generic browser automation service using browser-use"
dependencies = [
    "flask>=3.0.0",
    "requests>=2.31.0",
    "browser-use>=0.1.5",
    "python-dotenv>=1.0.0",
    "anthropic>=0.18.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
